Exception fixes for Python 3.10 (#21)
Fixed deprecation, fixed parameter type mismatch

+++ umlfri2/qtgui/mainwindow/tabs/startpageframe.py
@@ -29,8 +29,7 @@ class StartPageFrame(QWidget):
         
         viewport = painter.viewport()
         
-        rect = QRect(self.THICKNESS/2, self.THICKNESS/2, viewport.width() - self.THICKNESS,
-                     viewport.height() - self.THICKNESS)
+        rect = QRect(int(self.THICKNESS/2), int(self.THICKNESS/2), int(viewport.width() - self.THICKNESS), int(viewport.height() - self.THICKNESS))
         
         painter.drawRoundedRect(rect, 15, 15)
